\name{barplot.SCGLR}
\alias{barplot.SCGLR}
\title{barplot}
\usage{
  \method{barplot}{SCGLR} (height, ...)
}
\arguments{
  \item{height}{SCGLR result}

  \item{...}{unused}
}
\value{
  plot
}
\description{
  barplot
}
\seealso{
  For barplot application see examples in
  \code{\link{plot.SCGLR}}
}

