% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{barplot.SCGLRTHM}
\alias{barplot.SCGLRTHM}
\title{Barplot of percent of overall X variance captured by component}
\usage{
\method{barplot}{SCGLRTHM}(height, ...)
}
\arguments{
\item{height}{object of class 'SCGLRTHM', usually a result of running \code{\link{theme}}.}

\item{\dots}{optional arguments.}
}
\value{
an object of class ggplot.
}
\description{
Barplot of percent of overall X variance captured by component by theme
}
\keyword{internal}
