\name{critConvergence}
\alias{critConvergence}
\title{Auxiliary function tuning the convergence of scglr}
\usage{
  critConvergence(tol = 1e-06, maxit = 50)
}
\arguments{
  \item{tol}{positive convergence threshold}

  \item{maxit}{integer, maximum number of iterations}
}
\value{
  a list containing elements named as the arguments
}
\description{
  Function tuning the convergence of scglr
}

