% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.SCGLRTHM}
\alias{plot.SCGLRTHM}
\title{SCGLRTHM generic plot}
\usage{
\method{plot}{SCGLRTHM}(x, ...)
}
\arguments{
\item{x}{object of class 'SCGLRTHM', usually a result of running \code{\link{scglrTheme}}.}

\item{\dots}{see SCGLR plot method}
}
\value{
an object of class \code{ggplot}.
}
\description{
SCGLR generic plot for themes
}
