% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.SCGLR}
\alias{print.SCGLR}
\title{Print SCGLR object}
\usage{
\method{print}{SCGLR}(x, ...)
}
\arguments{
\item{x}{object of class 'SCGLR', usually a result of running \code{\link{scglr}}.}

\item{\dots}{Not used.}
}
\description{
Prints inertia per component and residual and null deviance for each Y.
}
\keyword{internal}
