% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{screeplot.SCGLR}
\alias{screeplot.SCGLR}
\title{Screeplot of percent of overall X variance captured by component}
\usage{
\method{screeplot}{SCGLR}(x, ...)
}
\arguments{
\item{x}{object of class 'SCGLR', usually a result of running \code{\link{scglr}}.}

\item{\dots}{optional arguments.}
}
\value{
an object of class ggplot.
}
\description{
Screeplot of percent of overall X variance captured by component
}
\seealso{
For screeplot application see examples in \code{\link{plot.SCGLR}}.
}
