% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{screeplot.SCGLRTHM}
\alias{screeplot.SCGLRTHM}
\title{Screeplot of percent of overall X variance captured by component}
\usage{
\method{screeplot}{SCGLRTHM}(x, ...)
}
\arguments{
\item{x}{object of class 'SCGLRTHM', usually a result of running \code{\link{theme}}.}

\item{\dots}{optional arguments.}
}
\value{
an object of class ggplot.
}
\description{
Screeplot of percent of overall X variance captured by component by theme
}
