\name{summary.SCGLR}
\alias{summary.SCGLR}
\title{Summary}
\usage{
  \method{summary}{SCGLR} (object, ..., digits = 3)
}
\arguments{
  \item{object}{a SCGLR object from scglr() function}

  \item{digits}{minimal number of significant digits, see
  print.default.}

  \item{...}{unused}
}
\description{
  Summary
}

