% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{barplot.SCGLR}
\alias{barplot.SCGLR}
\title{Barplot of percent of overall X variance captured by component}
\usage{
\method{barplot}{SCGLR}(height, ..., plane = NULL)
}
\arguments{
\item{height}{object of class 'SCGLR', usually a result of running \code{\link{scglr}}.}

\item{plane}{a size-2 vector (or comma separated string) indicating which components are plotted (eg: c(1,2) or "1,2").}

\item{\dots}{optional arguments.}
}
\value{
an object of class ggplot.
}
\description{
A custom plot for SCGLR objetcs
}
\seealso{
For barplot application see examples in \code{\link{plot.SCGLR}}.
}

