% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{barplot.SCGLR}
\alias{barplot.SCGLR}
\title{Barplot of percent of overall X variance captured by component}
\usage{
\method{barplot}{SCGLR}(height, ...)
}
\arguments{
\item{height}{object of class 'SCGLR', usually a result of running \code{\link{scglr}}.}

\item{\dots}{optional arguments.}
}
\value{
an object of class ggplot.
}
\description{
Barplot of percent of overall X variance captured by component
}
\seealso{
For barplot application see examples in \code{\link{plot.SCGLR}}.
}
\keyword{internal}
