% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{multivariateFormula}
\alias{multivariateFormula}
\title{Formula construction}
\usage{
multivariateFormula(namesY, namesX, namesAX = NULL)
}
\arguments{
\item{namesY}{a vector of character containing the names of the dependent variables.}

\item{namesX}{a vector of character containing the names of the covariates (X) involved
in the components.}

\item{namesAX}{a vector of character containing the names of the additional covariates.}
}
\value{
an object of class \code{Formula}.
}
\description{
Helper function for building multivariate scglr formula.

NOTE: Interactions involving factors are not allowed for now.
For interactions between two quantitative variables, use \code{I(x*y)} as usual.
}

