\name{pairs.SCGLR}
\alias{pairs.SCGLR}
\title{pairs}
\usage{
  \method{pairs}{SCGLR} (x, ..., nrow = NULL, ncol = NULL,
    plans = NULL)
}
\arguments{
  \item{x}{SCGLR result to plot}

  \item{...}{unused}

  \item{nrow}{number of rows of the grid layout}

  \item{ncol}{number of columns of the grid layout}

  \item{plans}{vector of integer to select components to
  use}
}
\value{
  plot
}
\description{
  pairs
}
\seealso{
  For pairs application see examples in
  \code{\link{plot.SCGLR}}
}

