% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot.SCGLRTHM}
\alias{plot.SCGLRTHM}
\title{SCGLRTHM generic plot}
\usage{
\method{plot}{SCGLRTHM}(x, ...)
}
\arguments{
\item{x}{an object from SCGLRTHM class.}

\item{\dots}{see SCGLR plot method}
}
\value{
an object of class \code{\link{ggplot}}.
}
\description{
SCGLR generic plot for themes
}
