% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.SCGLR}
\alias{summary.SCGLR}
\alias{print.summary.SCGLR}
\title{Summarizing SCGLR fits}
\usage{
\method{summary}{SCGLR}(object, ...)

\method{print}{summary.SCGLR}(x, digits = 3, ...)
}
\arguments{
\item{object}{an object of class "SCGLR", usually a result of a call to \code{\link{scglr}}.}

\item{\dots}{Not used.}

\item{x}{an object of class "summary.SCGLR", usually a result of a call to summary.SCGLR.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
an object of class "summary.SCGLR".
   \item{inertia}{inertia per component.}
   \item{deviance.null}{null deviance for each \eqn{Y_k}.}
   \item{deviance.residual}{residual deviance for each \eqn{Y_k}.}
   \item{rho}{squared correlations with numerical covariates.}
   \item{rho.pred}{squared correlations with linear predictors.}
   \item{coefficients}{contains the coefficients of the regression on the components.}
}
\description{
Summary method for class "SCGLR".
}
\keyword{internal}
